+++ app/app_test.go
@@ -9,16 +9,17 @@ import (
 	dbm "github.com/cosmos/cosmos-db"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	stktypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/stretchr/testify/require"
 
-	"github.com/babylonlabs-io/babylon/testutil/signer"
+	testsigner "github.com/babylonlabs-io/babylon/testutil/signer"
 	checkpointingtypes "github.com/babylonlabs-io/babylon/x/checkpointing/types"
 )
 
 func TestBabylonBlockedAddrs(t *testing.T) {
 	db := dbm.NewMemDB()
 
-	tbs, err := signer.SetupTestBlsSigner()
+	tbs, err := testsigner.SetupTestBlsSigner()
 	require.NoError(t, err)
 	blsSigner := checkpointingtypes.BlsSigner(tbs)
 
@@ -79,7 +80,7 @@ func TestGetMaccPerms(t *testing.T) {
 func TestUpgradeStateOnGenesis(t *testing.T) {
 	db := dbm.NewMemDB()
 
-	tbs, err := signer.SetupTestBlsSigner()
+	tbs, err := testsigner.SetupTestBlsSigner()
 	require.NoError(t, err)
 	blsSigner := checkpointingtypes.BlsSigner(tbs)
 
@@ -101,5 +102,34 @@ func TestUpgradeStateOnGenesis(t *testing.T) {
 		if i, ok := i.(module.HasConsensusVersion); ok {
 			require.Equal(t, vm[v], i.ConsensusVersion())
 		}
+	}
+}
+
+func TestStakingRouterDisabled(t *testing.T) {
+	db := dbm.NewMemDB()
+	tbs, _ := testsigner.SetupTestBlsSigner()
+	logger := log.NewTestLogger(t)
+
+	app := NewBabylonAppWithCustomOptions(t, false, tbs, SetupOptions{
+		Logger:             logger,
+		DB:                 db,
+		InvCheckPeriod:     0,
+		SkipUpgradeHeights: map[int64]bool{},
+		AppOpts:            TmpAppOptions(),
+	})
+
+	msgs := []sdk.Msg{
+		&stktypes.MsgCreateValidator{},
+		&stktypes.MsgBeginRedelegate{},
+		&stktypes.MsgCancelUnbondingDelegation{},
+		&stktypes.MsgDelegate{},
+		&stktypes.MsgEditValidator{},
+		&stktypes.MsgUndelegate{},
+		&stktypes.MsgUpdateParams{},
+	}
+
+	for _, msg := range msgs {
+		msgHandler := app.MsgServiceRouter().HandlerByTypeURL(sdk.MsgTypeURL(msg))
+		require.Nil(t, msgHandler)
 	}
 }