+++ cmd/babylond/cmd/genhelpers/cmd.go
@@ -18,8 +18,6 @@ func CmdGenHelpers(validator genutiltypes.MessageValidator) *cobra.Command {
 	}
 
 	cmd.AddCommand(
-		CmdCreateBls(),
-		CmdAddBls(validator),
 		CmdSetFp(),
 		CmdSetBtcDels(),
 		CmdSetBtcHeaders(),